+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 1.7.1-preview.0
+
+### Patch Changes
+
+- Updated dependencies [4060bc6]
+  - @dojoengine/torii-wasm@1.7.1-preview.0
+  - @dojoengine/grpc@1.7.1-preview.0
+  - @dojoengine/torii-client@1.7.1-preview.0
+  - @dojoengine/state@1.7.1-preview.0
+
 ## 1.7.0
 
 ### Minor Changes